Replace variables with macros (again)

+++ licharger.h
@@ -24,9 +24,9 @@ void stopWatchdog(void);
 		- Full battery is considered to be 4.2V, giving 853/1024
 		- Charge limit, i.e. level above which new charging will not start, is set at 4.0V, giving 812/1024
 	*/
-#define VLOWLVL 650 //Voltage where battery is considered depleted
-#define VHIGHLVL 853 //Charge to this voltage - 4.15V
-#define VCHARGELVL 812 //Voltage above which charging will not initiate - 4.0V
+#define VLOWLVL (uint16_t) 650 //Voltage where battery is considered depleted
+#define VHIGHLVL (uint16_t) 853 //Charge to this voltage - 4.15V
+#define VCHARGELVL (uint16_t) 812 //Voltage above which charging will not initiate - 4.0V
 
 //Booleans - Stored in state variable
 #define USBCONNECTED 1
